Run `allow-scripts` when restoring from cache (#11)
Packages based on the `metamask-module-template` run `allow-scripts` as
part of the Yarn install. When using `action-checkout-and-setup`, Yarn
is completely skipped however when restoring `node_modules` from cache,
meaning `allow-scripts` is skipped.

I've added a separate step to run `allow-scripts` if the cache was
restored, which ensures `allow-scripts` is run, either through Yarn, or
directly.